Initial stages of breaking out into components
The is a jumbled mess at the moment.  Old versions, changed in main app, partially completed sections etc.

This is an evolving thing, but I think now, I will get what I want in one app file and when that's running right, I'll re-visit this mess of ideas to break out components.

Idea of components to allow me to run smaller sections or partial imports etc.