Remove unnecessary async/s & await/s
Issue with files attempting to be written before directory creation had completed.

Removed any (what I determined to be) unnecessary async and awaits.

Also changed both the directory writes to the Synchronous version in an attempt to make sure they are created before returning.

Seems to have solved the problem, though I've currently only tested it twice on small batch (2 katas).  So not sure if a larger batch would cause sync issues or not at this stage.